requirements.txt
# Chain of Thought Text Processing Pipeline
# Core dependencies

groq>=0.5.0                    # Groq LLM API client
python-dotenv>=1.0.0           # Environment variable management
contractions>=0.1.73           # Expand text contractions
langdetect>=1.0.9              # Language detection
markdownify>=0.11.0            # HTML to Markdown conversion
tiktoken>=0.5.0                # Token counting for OpenAI-compatible models
beautifulsoup4>=4.12.0         # HTML parsing
lxml>=5.0.0                    # Fast HTML/XML parser
regex>=2023.0.0                # Advanced regex operations


# Testing
pytest>=7.4.0

# API & Frontend
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
pydantic>=2.6.0
